feat: 카트로직 분리
- 분리 중 테스트케이스 실패 이슈
quantity-input.value 값을 state를 보고업데이트했는데,
state업데이트시 바로 렌더링되도록해서 quantity-input.value값이 렌더링 후 업데이트 되는 문제가 있었음. quantity-input.value 값을 업데이트하는 함수도 state 변경 subsscribe하여 업데이트 되도록 수정
